TASK 2+3: functionality of djikstra's algorithm, fixing broken code

+++ src/data/Edge.java
@@ -5,10 +5,10 @@ public class Edge {
     // variables to assign departure, arrival, and time taken from CSV list for each station
     public final String departure;
     public final String arrival;
-    public final int time;
+    public final double time;
 
     // function assigning passed-in values from CSV file to related variables
-    public Edge(String departure, String arrival, int time) {
+    public Edge(String departure, String arrival, double time) {
         this.departure = departure;
         this.arrival = arrival;
         this.time = time;
