+++ src/model/RouteCalc.java
@@ -3,9 +3,9 @@ import java.util.List;
 
 public class RouteCalc {
     public final List<String> path;
-    public final int total;
+    public final double total;
 
-    public RouteCalc(List<String> path, int total) {
+    public RouteCalc(List<String> path, double total) {
         this.path = path;
         this.total = total;
     }
